requirements.txt
djoser
djangorestframework-simplejwt==4.7.2
